requirements.txt
torch
transformers
datasets
torchmetrics==1.4.0
accelerate
timm
albumentations==1.4.6
pycocotools
Pillow>=10.0.0
numpy>=1.20.0